feat(ts-interface-generator): generate interface in a way that requires no named export anymore
The generated interface is now created in a way that the original control file no longer needs an
additional named export of the control. This is achieved by exporting the generated interface as
default export. Right now, however, it appears like to avoid TypeScript errors within the interface
this requires a duplicate definition of the interface - once without being default export. It is
being investigated how this can be avoided.