fix(ts-interface-generator): fix crash with parent w/o settings type
To generate constructor signatures, a settings type needs to be created
which inherits from the
parent class' settings type and adds the properties, aggregations etc.
of the current class. This
rightly leads to an error. But this should not lead to an error when the
class does not add any API
items at all.

related to #346