fix(ts-interface-generator): handle gracefully the case when metadata is not properly initialized
When the user accidentally writes "metadata: ..." instead of "metadata = ...", then this is
syntactically valid, but it declares the type instead of actually defining the metadata of the
control. Prior to this fix, this led to a crash in the generator, now a hopefully helpful warning is
displayed and the generator ignores this class.

fix #369