fix(ts-interface-generator): remove EventProvider from list of ManagedObject sub-classes
EventProvider was wrongly added as one of the parent classes which signals that an interface needs
to be generated. But accessors are only created for descendants of ManagedObject and EventProvider
is the super-class, not a sub-class. This would lead to issues because assumptions done on the base
that a ManagedObject is present were wrong.

fix #374