fix(ts-interface-generator): do not generate interfaces for classes outside the project
Previously, the path of all TS files considered by the compiler was checked for "node_modules" and
"@types" to exclude those from the list which are in external libraries. Only the files not being in
such folder were actually used for generating interfaces. However, in a setup with a path mapping,
the written, not the resolved path was available during the filtering, so accidentally the generator
handled files it should not handle. This fix adds a check whether from compiler perspective the file
is "from an external library".