chore(ts-interface-generator): update dependencies + adapt
This is a pure dependency update, but also an adaptation to the overall
repo dependency update, where a new major version of prettier was
applied, which causes lots of formatting changes (mostly adding trailing
commas) in this change.
Also, TypeScript was brought in sync to the other parts of the repo.