feat: add dependenciesTypePackagesForCheck option
Basically this adds the "types:" option in the tsc run during the
compile check.
This was not needed in the standard UI5 build environment, but when run
elsewhere:

1. the OpenUI5 types need to be added as dependencies and doing this via
npm package name is cleaner than pointing into the node_modules folder
using the dependencyDTSFilesForCheck option and

2. tsc may automatically pick up other type packages which happen to be
around and report errors or missing dependencies inside them. Setting
"types" means only *these* types are considered.