chore: update dependencies
Using the new eslint and TypeScript version requires the removal of a
type cast which is now unnecessary, as TypeScript 5.5 recognizes that
after .filter() with a type check all remaining items have this type.

In current vscode, this requires the manual selection of the workspace
version of TypeScript, as vscode itself still uses version 5.4.