chore: convert from dep5 to REUSE.toml
Also, remove the license-copying, which was basically wrong as 1.) REUSE
info is only relevant in the repository, not in the npm package, and 2.)
even if it was relevant, copying from the root to all subfolders would
have broken references and added license infos which are not even
relevant for each subpackage. If only one subpackage would have been
under a different license or have contained single files under a
different license, then all packages would have received a copy of that
license.
Instead, each published npm package gets a copy of the license.