fix(dts-generator): fix more derived type names for module:* classes
When a managed object class has a module:* name, the names are now
correctly split into prefix and base name when calculating the name
of the EventParameters and EventType types.

The implementation of the calculations is now shared with the
calculation implemented in b6c0fe52c1603042439df035b5ba8f5439b732f8.